Clean up frequenz.sdk.actor public API
Also make some minor improvements to the documentation.
