+++ src/frequenz/sdk/actor/_data_sourcing/__init__.py
@@ -4,10 +4,9 @@
 """The DataSourcingActor."""
 
 from .data_sourcing import DataSourcingActor
-from .types import ComponentMetricId, ComponentMetricRequest
+from .microgrid_api_source import ComponentMetricRequest
 
 __all__ = [
-    "DataSourcingActor",
-    "ComponentMetricId",
     "ComponentMetricRequest",
+    "DataSourcingActor",
 ]
