+++ src/frequenz/sdk/actor/_data_sourcing/microgrid_api_source.py
@@ -5,6 +5,8 @@
 
 import asyncio
 import logging
+from dataclasses import dataclass
+from datetime import datetime
 from typing import Any, Callable, Dict, List, Optional, Tuple
 
 from frequenz.channels import Receiver, Sender
@@ -12,14 +14,54 @@ from frequenz.channels import Receiver, Sender
 from ...microgrid import (
     BatteryData,
     ComponentCategory,
+    ComponentMetricId,
     EVChargerData,
     InverterData,
     MeterData,
     microgrid_api,
 )
 from ...timeseries import Sample
-from .. import ChannelRegistry
-from .types import ComponentMetricId, ComponentMetricRequest
+from .._channel_registry import ChannelRegistry
+
+
+@dataclass
+class ComponentMetricRequest:
+    """A request object to start streaming a metric for a component."""
+
+    namespace: str
+    """The namespace that this request belongs to.
+
+    Metric requests with a shared namespace enable the reuse of channels within
+    that namespace.
+
+    If for example, an actor making a multiple requests, uses the name of the
+    actor as the namespace, then requests from the actor will get reused when
+    possible.
+    """
+
+    component_id: int
+    """The ID of the requested component."""
+
+    metric_id: ComponentMetricId
+    """The ID of the requested component's metric."""
+
+    start_time: Optional[datetime]
+    """The start time from which data is required.
+
+    When None, we will stream only live data.
+    """
+
+    def get_channel_name(self) -> str:
+        """Return a channel name constructed from Self.
+
+        This channel name can be used by the sending side and receiving sides to
+        identify the right channel from the ChannelRegistry.
+
+        Returns:
+            A string denoting a channel name.
+        """
+        return f"{self.component_id}::{self.metric_id.name}::{self.start_time}::{self.namespace}"
+
 
 _MeterDataMethods: Dict[ComponentMetricId, Callable[[MeterData], float]] = {
     ComponentMetricId.ACTIVE_POWER: lambda msg: msg.active_power,
