+++ src/frequenz/sdk/data_ingestion/microgrid_data.py
@@ -18,7 +18,7 @@ from typing import Any, Dict, List, Optional, Set
 from frequenz.channels import Broadcast, Receiver, Sender
 from frequenz.channels.util import Merge, Select
 
-from ..actor.decorator import actor
+from ..actor import actor
 from ..configs import Config
 from ..data_handling.time_series import TimeSeriesEntry
 from ..microgrid.client import MicrogridApiClient
