+++ src/frequenz/sdk/microgrid/__init__.py
@@ -9,7 +9,7 @@ for monitoring and adjusting the state of a microgrid.
 
 from . import microgrid_api
 from .client import MicrogridApiClient
-from .component import Component, ComponentCategory
+from .component import Component, ComponentCategory, ComponentMetricId
 from .component_data import BatteryData, EVChargerData, InverterData, MeterData
 from .connection import Connection
 from .graph import ComponentGraph
@@ -19,6 +19,7 @@ __all__ = [
     "Component",
     "ComponentGraph",
     "ComponentCategory",
+    "ComponentMetricId",
     "Connection",
     "BatteryData",
     "EVChargerData",
